+++ frontend/src/pages/Recipients/index.js
@@ -6,6 +6,7 @@ import { useHistory } from 'react-router-dom';
 import Table from '../../components/Table';
 import Input from '../../components/Input';
 import RegisterButton from '../../components/RegisterButton';
+import NothingHere from '../../components/NothingHere';
 import DialogContent from './DialogContent';
 
 import goToRegister from '../../utils/goToRegister';
@@ -15,44 +16,47 @@ import api from '../../services/api';
 function Recipients() {
   const history = useHistory();
 
+  const [searchValue, setSearchValue] = useState('');
+  const [loading, setLoading] = useState(false);
+
   const [recipients, setRecipients] = useState([]);
+  const [hasRecipients, setHasRecipients] = useState(true);
+
   const [tableContent, setTableContent] = useState({
     headItems: ['ID', 'Nome', 'Endereço', 'Ações'],
   });
 
-  const wait = (ms) =>
-    new Promise((resolve) => {
-      setTimeout(() => {
-        resolve('pesquisa completa');
-      }, ms);
-    });
-
-  async function getRecipient() {
-    await wait(2000);
-  }
-
   async function getRecipients() {
     try {
-      const response = await api.get('/recipients');
+      setLoading(true);
+
+      const response = await api.get('/recipients', {
+        params: { q: searchValue },
+      });
       setRecipients(response.data);
+      setLoading(false);
+
+      setHasRecipients(!!response.data[0]);
     } catch (err) {
       toast.error('Falha ao listar destinatários.');
     }
   }
 
   async function handleDelete(id) {
-    try {
-      await api.delete(`/recipients/${id}`, { params: { destroy: 'true' } });
-      toast.success('Destinatário deletado com sucesso.');
-      getRecipients();
-    } catch (err) {
-      toast.error('Falha ao deletar destinatário.');
+    if (window.confirm('Deletar registro permanentemente?')) {
+      try {
+        await api.delete(`/recipients/${id}`, { params: { destroy: 'true' } });
+        toast.success('Destinatário deletado com sucesso.');
+        getRecipients();
+      } catch (err) {
+        toast.error('Falha ao deletar destinatário.');
+      }
     }
   }
 
   useEffect(() => {
     getRecipients();
-  }, []);
+  }, [searchValue]);
 
   useEffect(() => {
     const rows = recipients.map((recipient) => {
@@ -81,20 +85,32 @@ function Recipients() {
             type="text"
             placeholder="Buscar por destinatários"
             typeName="search"
-            onSearch={getRecipient}
+            onSearch={getRecipients}
+            loading={loading}
+            onChange={(evt) => setSearchValue(evt.target.value)}
           />
 
           <RegisterButton onClick={() => goToRegister(history)} />
         </div>
       </header>
 
-      <Table
-        headItems={tableContent.headItems}
-        bodyRows={tableContent.rows}
-        dialog={{ Component: DialogContent, title: 'Destinatário' }}
-        category="recipients"
-        handleDelete={handleDelete}
-      />
+      {hasRecipients ? (
+        <Table
+          headItems={tableContent.headItems}
+          bodyRows={tableContent.rows}
+          dialog={{ Component: DialogContent, title: 'Destinatário' }}
+          category="recipients"
+          handleDelete={handleDelete}
+          loading={loading}
+          optionsList={[
+            { key: 'view', label: 'Visualizar' },
+            { key: 'edit', label: 'Editar' },
+            { key: 'delete', label: 'Excluir' },
+          ]}
+        />
+      ) : (
+        <NothingHere>Nenhum destinatário foi encontrado.</NothingHere>
+      )}
     </>
   );
 }
